fastapi-backend/requirements.txt
fastapi
torch
pillow
uvicorn
torchvision
